# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "API for easy study new things from school",
        "version": "0.3.4",
        "title": "Learn and test knowledge - API"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/quiz/schoolSubjects/subjects": {
            "get": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Interface can take all subjects",
                "responses": {
                    "200": {
                        "description": "List of school subjects"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "501": {
                        "description": "System error - subjects doesn't exist"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/chapters": {
            "get": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Interface can take all chapters",
                "responses": {
                    "200": {
                        "description": "List of chapters"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/topics": {
            "get": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Interface can take all topics",
                "responses": {
                    "200": {
                        "description": "List of topics"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/chapter": {
            "post": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Teacher can add new chapter",
                "requestBody": {
                    "description": "Chapter data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_subject": {
                                        "type": "string"
                                    },
                                    "name_of_chapter": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New chapter added successfully"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/topic": {
            "post": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Teacher can add new topic",
                "requestBody": {
                    "description": "Topic data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_chapter": {
                                        "type": "string"
                                    },
                                    "name_of_topic": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New topic added successfully"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/remove-chapter": {
            "put": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Teacher can delete chapter",
                "requestBody": {
                    "description": "Chapter data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_chapter": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/quiz/schoolSubjects/remove-topic": {
            "put": {
                "tags": [
                    "School subjects"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Teacher can delete topic",
                "requestBody": {
                    "description": "Topic data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name_of_topic": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully completed"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    }
}